backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
google-generativeai==0.7.2
Pillow==10.3.0
aiofiles==23.2.1
websockets==12.0
pydantic==2.7.1
pydantic-settings==2.2.1
cloudinary==1.40.0
redis==5.0.4
celery==5.4.0
pytest==8.2.0
pytest-asyncio==0.23.7
email-validator==2.1.1
asyncpg==0.29.0
